requirements.txt
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
postgres==2.2.2
psycopg2-binary==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
